// Neptune.Web/src/app/shared/generated/model/treatment-bmp-assessment-observation-type-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPAssessmentObservationTypeUpsertDto { 
    TreatmentBMPAssessmentObservationTypeName: string;
    ObservationTypeSpecificationID: number;
    TreatmentBMPAssessmentObservationTypeSchema?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPAssessmentObservationTypeUpsertDtoForm { 
    TreatmentBMPAssessmentObservationTypeName: FormControl<string>;
    ObservationTypeSpecificationID: FormControl<number>;
    TreatmentBMPAssessmentObservationTypeSchema?: FormControl<string>;
}

export class TreatmentBMPAssessmentObservationTypeUpsertDtoFormControls { 
    public static TreatmentBMPAssessmentObservationTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
                Validators.maxLength(100),
            ],
        }
    );
    public static ObservationTypeSpecificationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static TreatmentBMPAssessmentObservationTypeSchema = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}